Break out utils/shared module
