--- src/utils/index.js
@@ -1,13 +1,0 @@
-/**
- * Convert str to title case (every word will be capitalized)
- * @param {String} str
- * @return {String}
- */
-function titleCase (str) {
-  return str
-    .toLowerCase()
-    .replace(/(\-|_)/g, ' ')
-    .replace(/\w\S*/g, word => word.charAt(0).toUpperCase() + word.substr(1));
-}
-
-export { titleCase };
